+++ ChangeLog.md
@@ -13,6 +13,8 @@
 - `optparse-applicative` lower bound tightened to 0.17.1
 - `Brassica.SoundChange.Apply.Internal.applyOnce` now returns a `RuleStatus` value with more detailed information about the rule application, which is now used by `Brassica.SoundChange.Apply.Internal.setupForNextApplication`
 - `Wildcard` and `Kleene` no longer have `OneOf 'Target 'Env` constraint
+- `OneOf` type family is no longer used and has been removed
+- `Target` and `Environment` `LexemeType`s have been unified as `Matched`
 
 ## v0.2.0